+++ frontend/src/static/js/components/media-actions/MediaShareButton.jsx
@@ -1,51 +1,10 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { usePopup } from '../../utils/hooks/';
-import { CircleIconButton, MaterialIcon, NavigationContentApp, PopupMain } from '../_shared/';
-import { MediaShareEmbed } from './MediaShareEmbed';
+import { CircleIconButton, MaterialIcon, PopupMain } from '../_shared/';
 import { MediaShareOptions } from './MediaShareOptions';
 
-function mediaSharePopupPages() {
-  return {
-    shareOptions: (
-      <div className="popup-fullscreen">
-        <PopupMain>
-          <span className="popup-fullscreen-overlay"></span>
-          <MediaShareOptions />
-        </PopupMain>
-      </div>
-    ),
-  };
-}
-
-function videoSharePopupPages(onTriggerPopupClose) {
-  return {
-    ...mediaSharePopupPages(),
-    shareEmbed: (
-      <div className="popup-fullscreen share-embed-popup">
-        <PopupMain>
-          <span className="popup-fullscreen-overlay"></span>
-          <MediaShareEmbed triggerPopupClose={onTriggerPopupClose} />
-        </PopupMain>
-      </div>
-    ),
-  };
-}
-
-export function MediaShareButton(props) {
+export function MediaShareButton() {
   const [popupContentRef, PopupContent, PopupTrigger] = usePopup();
-
-  const [popupCurrentPage, setPopupCurrentPage] = useState('shareOptions');
-
-  function triggerPopupClose() {
-    popupContentRef.current.toggle();
-  }
-
-  function onPopupPageChange(newPage) {
-    setPopupCurrentPage(newPage);
-  }
-  function onPopupHide() {
-    setPopupCurrentPage('shareOptions');
-  }
 
   return (
     <div className="share">
@@ -58,15 +17,13 @@ export function MediaShareButton(props) {
         </button>
       </PopupTrigger>
 
-      <PopupContent contentRef={popupContentRef} hideCallback={onPopupHide}>
-        <NavigationContentApp
-          initPage={popupCurrentPage}
-          pageChangeSelector={'.change-page'}
-          pageIdSelectorAttr={'data-page-id'}
-          pages={props.isVideo ? videoSharePopupPages(triggerPopupClose) : mediaSharePopupPages()}
-          focusFirstItemOnPageChange={false}
-          pageChangeCallback={onPopupPageChange}
-        />
+      <PopupContent contentRef={popupContentRef}>
+        <div className="popup-fullscreen">
+          <PopupMain>
+            <span className="popup-fullscreen-overlay"></span>
+            <MediaShareOptions />
+          </PopupMain>
+        </div>
       </PopupContent>
     </div>
   );