+++ frontend/src/static/js/components/media-actions/index.js
@@ -2,7 +2,6 @@ export * from './MediaMoreOptionsIcon.jsx';
 export * from './MediaSaveButton.jsx';
 export * from './MediaTechniqueButton.jsx';
 export * from './MediaShareButton.jsx';
-export * from './MediaShareEmbed.jsx';
 export * from './MediaShareOptions.jsx';
 export * from './OtherMediaDownloadLink.jsx';
 export * from './VideoMediaDownloadLink.jsx';