+++ frontend/src/static/js/utils/settings/media.js
@@ -29,16 +29,11 @@ export function init(item, shareOptions) {
 
   if (void 0 !== shareOptions) {
     const validShareOptions = [
-      'embed',
       'fb',
       'tw',
       'whatsapp',
       'telegram',
       'reddit',
-      'tumblr',
-      'vk',
-      'pinterest',
-      'mix',
       'linkedin',
       'email',
     ];